Fixed a bunch of type issues, primarily iteration mismatches

+++ src/pcap_device.h
@@ -28,12 +28,12 @@ struct Device {
 
 struct Device *CreateDevice(const char *name);
 uint8_t FreeDevice(struct Device *device);
-uint8_t StartDevice(struct Device *device);
+int StartDevice(struct Device *device);
 uint8_t StopDevice(struct Device *device);
 
 int AddAddress(struct Device *device, const char *address, const int type);
 int AddressExists(const struct Device *device, const char *address, const int type);
-int GetNoAddresses(const struct Device *device);
+size_t GetNoAddresses(const struct Device *device);
 int RemoveAddress(struct Device *device, const char *address);
 void RemoveAllAddresses(struct Device *device);
 int SetAllAddresses(struct Device *device);
